# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "EvalHarness API",
  "description": "Runs agentic tool-use scenarios against any LiteLLM-supported model and grades them deterministically.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/attempts/{attempt_id}": {
   "get": {
    "tags": [
     "attempts"
    ],
    "summary": "Full trace: turns, ordered events and assertion results",
    "operationId": "getAttempt",
    "parameters": [
     {
      "name": "attempt_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Attempt Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttemptTrace"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/compare": {
   "get": {
    "tags": [
     "leaderboard"
    ],
    "summary": "Side-by-side summary, per-scenario diffs and assertions that flipped",
    "operationId": "compareRuns",
    "parameters": [
     {
      "name": "run_a",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Baseline run id.",
       "title": "Run A"
      },
      "description": "Baseline run id."
     },
     {
      "name": "run_b",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Run to compare against the baseline.",
       "title": "Run B"
      },
      "description": "Run to compare against the baseline."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CompareResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "AssertionFlip": {
    "properties": {
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "assertion_id": {
      "type": "string",
      "title": "Assertion Id"
     },
     "type": {
      "type": "string",
      "title": "Type",
      "default": ""
     },
     "axis": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Axis"
       },
       {
        "type": "null"
       }
      ]
     },
     "severity": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Severity"
       },
       {
        "type": "null"
       }
      ]
     },
     "pass_rate_a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass Rate A"
     },
     "pass_rate_b": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass Rate B"
     },
     "direction": {
      "type": "string",
      "enum": [
       "fixed",
       "broken",
       "improved",
       "regressed",
       "added",
       "removed"
      ],
      "title": "Direction"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario_id",
     "assertion_id",
     "direction"
    ],
    "title": "AssertionFlip",
    "description": "An assertion whose outcome moved between the two runs."
   },
   "AssertionResultOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "attempt_id": {
      "type": "string",
      "title": "Attempt Id"
     },
     "turn_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Turn Index"
     },
     "assertion_id": {
      "type": "string",
      "title": "Assertion Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "axis": {
      "$ref": "#/components/schemas/Axis"
     },
     "severity": {
      "$ref": "#/components/schemas/Severity"
     },
     "passed": {
      "type": "boolean",
      "title": "Passed"
     },
     "reason": {
      "type": "string",
      "title": "Reason",
      "default": ""
     },
     "details": {
      "additionalProperties": {
       "$ref": "#/components/schemas/JsonValue"
      },
      "type": "object",
      "title": "Details"
     },
     "non_deterministic": {
      "type": "boolean",
      "title": "Non Deterministic",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "attempt_id",
     "assertion_id",
     "type",
     "axis",
     "severity",
     "passed"
    ],
    "title": "AssertionResultOut"
   },
   "AttemptStatus": {
    "type": "string",
    "enum": [
     "queued",
     "running",
     "completed",
     "failed",
     "cancelled"
    ],
    "title": "AttemptStatus"
   },
   "AttemptSummary": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "repetition": {
      "type": "integer",
      "title": "Repetition"
     },
     "status": {
      "$ref": "#/components/schemas/AttemptStatus"
     },
     "passed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "critical_failure": {
      "type": "boolean",
      "title": "Critical Failure",
      "default": false
     },
     "axis_scores": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd"
     },
     "input_tokens": {
      "type": "integer",
      "title": "Input Tokens",
      "default": 0
     },
     "output_tokens": {
      "type": "integer",
      "title": "Output Tokens",
      "default": 0
     },
     "duration_ms": {
      "type": "integer",
      "title": "Duration Ms",
      "default": 0
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "run_id",
     "scenario_id",
     "repetition",
     "status"
    ],
    "title": "AttemptSummary"
   },
   "AttemptTrace": {
    "properties": {
     "attempt": {
      "$ref": "#/components/schemas/AttemptSummary"
     },
     "run": {
      "$ref": "#/components/schemas/Run"
     },
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "scenario_title": {
      "type": "string",
      "title": "Scenario Title",
      "default": ""
     },
     "config_hash": {
      "type": "string",
      "title": "Config Hash",
      "default": ""
     },
     "turns": {
      "items": {
       "$ref": "#/components/schemas/TurnOut"
      },
      "type": "array",
      "title": "Turns"
     },
     "events": {
      "items": {
       "$ref": "#/components/schemas/EventOut"
      },
      "type": "array",
      "title": "Events"
     },
     "assertion_results": {
      "items": {
       "$ref": "#/components/schemas/AssertionResultOut"
      },
      "type": "array",
      "title": "Assertion Results"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "attempt",
     "run",
     "scenario_id"
    ],
    "title": "AttemptTrace",
    "description": "Everything the trace viewer needs for one attempt (SPEC 10.1)."
   },
   "Axis": {
    "type": "string",
    "enum": [
     "selection",
     "arguments",
     "ordering",
     "restraint",
     "recovery",
     "state",
     "clarification",
     "safety"
    ],
    "title": "Axis",
    "description": "Capability dimension an assertion measures. Fixed set -- SPEC 4.3."
   },
...
   "CompareResponse": {
    "properties": {
     "run_a": {
      "$ref": "#/components/schemas/Run"
     },
     "run_b": {
      "$ref": "#/components/schemas/Run"
     },
     "scenarios": {
      "items": {
       "$ref": "#/components/schemas/ScenarioDiff"
      },
      "type": "array",
      "title": "Scenarios"
     },
     "flipped_assertions": {
      "items": {
       "$ref": "#/components/schemas/AssertionFlip"
      },
      "type": "array",
      "title": "Flipped Assertions"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "run_a",
     "run_b"
    ],
    "title": "CompareResponse"
   },
...
   "EventOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "attempt_id": {
      "type": "string",
      "title": "Attempt Id"
     },
     "turn_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Turn Index"
     },
     "seq": {
      "type": "integer",
      "title": "Seq"
     },
     "type": {
      "$ref": "#/components/schemas/EventType"
     },
     "payload": {
      "additionalProperties": {
       "$ref": "#/components/schemas/JsonValue"
      },
      "type": "object",
      "title": "Payload"
     },
     "latency_ms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency Ms"
     },
     "input_tokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Input Tokens"
     },
     "output_tokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Output Tokens"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "attempt_id",
     "seq",
     "type",
     "created_at"
    ],
    "title": "EventOut"
   },
   "EventType": {
    "type": "string",
    "enum": [
     "system",
     "user_message",
     "model_request",
     "model_response",
     "tool_call",
     "tool_result",
     "fault",
     "retry",
     "limit_exceeded",
     "error"
    ],
    "title": "EventType",
    "description": "Ordered audit-trail event kinds (SPEC 8.3)."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "JsonValue": {},
...
   "Run": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "status": {
      "$ref": "#/components/schemas/RunStatus"
     },
     "model_key": {
      "type": "string",
      "title": "Model Key"
     },
     "litellm_model": {
      "type": "string",
      "title": "Litellm Model"
     },
     "params": {
      "additionalProperties": {
       "$ref": "#/components/schemas/JsonValue"
      },
      "type": "object",
      "title": "Params"
     },
     "k": {
      "type": "integer",
      "title": "K"
     },
     "scenario_ids": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Scenario Ids"
     },
     "config_hashes": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Config Hashes"
     },
     "git_commit": {
      "type": "string",
      "title": "Git Commit",
      "default": "unknown"
     },
     "harness_version": {
      "type": "string",
      "title": "Harness Version",
      "default": "0.0.0"
     },
     "summary": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunSummary"
       },
       {
        "type": "null"
       }
      ]
     },
     "attempt_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Attempt Counts",
      "description": "Attempt count by status."
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "created_at",
     "status",
     "model_key",
     "litellm_model",
     "k"
    ],
    "title": "Run",
    "description": "A run without its (large) scenario snapshots."
   },
...
   "RunStatus": {
    "type": "string",
    "enum": [
     "queued",
     "running",
     "completed",
     "failed",
     "cancelled"
    ],
    "title": "RunStatus"
   },
   "RunSummary": {
    "properties": {
     "k": {
      "type": "integer",
      "title": "K",
      "default": 1
     },
     "total_attempts": {
      "type": "integer",
      "title": "Total Attempts",
      "default": 0
     },
     "completed_attempts": {
      "type": "integer",
      "title": "Completed Attempts",
      "default": 0
     },
     "passed_attempts": {
      "type": "integer",
      "title": "Passed Attempts",
      "default": 0
     },
     "failed_attempts": {
      "type": "integer",
      "title": "Failed Attempts",
      "default": 0
     },
     "cancelled_attempts": {
      "type": "integer",
      "title": "Cancelled Attempts",
      "default": 0
     },
     "errored_attempts": {
      "type": "integer",
      "title": "Errored Attempts",
      "default": 0
     },
     "pass_at_1": {
      "type": "number",
      "title": "Pass At 1",
      "default": 0.0
     },
     "pass_hat_k": {
      "type": "number",
      "title": "Pass Hat K",
      "default": 0.0
     },
     "axis_scores": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd",
      "description": "Null when any attempt's cost is unknown; never guessed."
     },
     "latency_p50_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms"
     },
     "latency_p95_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P95 Ms"
     },
     "model_calls": {
      "type": "integer",
      "title": "Model Calls",
      "default": 0
     },
     "input_tokens": {
      "type": "integer",
      "title": "Input Tokens",
      "default": 0
     },
     "output_tokens": {
      "type": "integer",
      "title": "Output Tokens",
      "default": 0
     },
     "mean_steps_per_turn": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mean Steps Per Turn"
     },
     "per_scenario": {
      "items": {
       "$ref": "#/components/schemas/ScenarioStats"
      },
      "type": "array",
      "title": "Per Scenario"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "RunSummary",
    "description": "The ``runs.summary`` jsonb blob, computed by ``finalize_run`` (SPEC 6.3)."
   },
...
   "ScenarioDiff": {
    "properties": {
     "scenario_id": {
      "type": "string",
      "title": "Scenario Id"
     },
     "config_hash_a": {
      "type": "string",
      "title": "Config Hash A",
      "default": ""
     },
     "config_hash_b": {
      "type": "string",
      "title": "Config Hash B",
      "default": ""
     },
     "config_changed": {
      "type": "boolean",
      "title": "Config Changed",
      "default": false
     },
     "pass_at_1_a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass At 1 A"
     },
     "pass_at_1_b": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass At 1 B"
     },
     "pass_at_1_delta": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass At 1 Delta"
     },
     "pass_hat_k_a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass Hat K A"
     },
     "pass_hat_k_b": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass Hat K B"
     },
     "pass_hat_k_delta": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pass Hat K Delta"
     },
     "axis_scores_a": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores A"
     },
     "axis_scores_b": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Scores B"
     },
     "axis_deltas": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Axis Deltas"
     },
     "cost_usd_a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd A"
     },
     "cost_usd_b": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd B"
     },
     "cost_usd_delta": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd Delta"
     },
     "latency_p50_ms_a": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms A"
     },
     "latency_p50_ms_b": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms B"
     },
     "latency_p50_ms_delta": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latency P50 Ms Delta"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "scenario_id"
    ],
    "title": "ScenarioDiff"
   },
...
   "TurnOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "attempt_id": {
      "type": "string",
      "title": "Attempt Id"
     },
     "index": {
      "type": "integer",
      "title": "Index"
     },
     "user_message": {
      "type": "string",
      "title": "User Message"
     },
     "final_response": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Response"
     },
     "passed": {
      "type": "boolean",
      "title": "Passed",
      "default": true
     },
     "limit_exceeded": {
      "type": "boolean",
      "title": "Limit Exceeded",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "attempt_id",
     "index",
     "user_message"
    ],
    "title": "TurnOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}